secured remove_user

+++ user/remove_user.php
@@ -1,9 +1,10 @@
 <?php
 
     include_once("../top.php");
-
-    $user_id = $_POST['delete_id'];
-    $user = new User($user_id);
-    $username = $user->getUsername();
-    $user->delete();
-    chatline("deletes user '$username'");
+    if($access == 0){
+        $user_id = $_POST['delete_id'];
+        $user = new User($user_id);
+        $username = $user->getUsername();
+        $user->delete();
+        chatline("deletes user '$username'");
+    }
